refactor: :bento: move `/utils` folder to `/lib`
prefer using the lib folder to include files used throughout the application; remove alias to the utils path; access files through `$lib/utils/*`
