+++ svelte.config.js
@@ -1,5 +1,4 @@
 import adapter from '@sveltejs/adapter-auto';
-import { resolve } from 'path';
 import { mdsvex } from 'mdsvex';
 import mdsvexConfig from './mdsvex.config.js';
 
@@ -8,14 +7,7 @@ const config = {
 	extensions: ['.svelte', ...mdsvexConfig.extensions],
 	preprocess: mdsvex(mdsvexConfig),
 	kit: {
-		adapter: adapter(),
-		vite: {
-			resolve: {
-				alias: {
-					$utils: resolve('./src/utils')
-				}
-			}
-		}
+		adapter: adapter()
 	}
 };
